// modul-2/day-7/global-state-app/src/types/content.ts
export interface TextContent {
  id: {
    title: string;
    langLabel: string;
    btnLang: string;
    welcome?: string;
    pleaseLogin: string;
    btnLogin: string;
    btnLogout: string;
    statusLogin: string;
    userActive: string;
    guest: string;
  };
  en: {
    title: string;
    langLabel: string;
    btnLang: string;
    welcome?: string;
    pleaseLogin: string;
    btnLogin: string;
    btnLogout: string;
    statusLogin: string;
    userActive: string;
    guest: string;
  };
}